+++ hasklepias-core/src/Cohort/AssessmentIntervals.hs
@@ -11,7 +11,6 @@ Maintainer  : bsaul@novisci.com
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
--- {-# LANGUAGE Safe #-}
 
 module Cohort.AssessmentIntervals
   (